Nelmio api

+++ src/Controller/UserController.php
@@ -5,8 +5,8 @@ namespace App\Controller;
 use App\Entity\UserClient;
 use App\Repository\UserClientRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Nelmio\ApiDocBundle\Annotation\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -19,9 +19,18 @@ use Nelmio\ApiDocBundle\Annotation\Model;
 class UserController extends AbstractController
 {
     /**
-     * Récuprération des utilisateurs en lien avec le client
+     * Récupération des utilisateurs en lien avec le client
      * @Route("/api/users", name="api_users_index", methods="GET")
      * @TAG\Tag(name="Users")
+     * @TAG\Response(
+     *     response=200,
+     *     description="Return l'ensemble des utilisateurs liée à un client",
+     *     @Model(type=UserClient::class, groups={"user:read"})
+     * )
+     * @TAG\Response(
+     *     response=401,
+     *     description="JWT Token not found | Expired JWT Token"
+     * )
      */
     public function getUsers(UserClientRepository $UserClientRepository)
     {
@@ -60,11 +69,16 @@ class UserController extends AbstractController
      *     description="User not found."
      * )
      * @TAG\Tag(name="Users")
-     * @Security(name="Bearer")
      */
-    public function getUserDetails(UserClient $user, UserClientRepository $UserClientRepository):Response
+    public function getUserDetail(UserClientRepository $UserClientRepository, string $id):Response
     {
-        $user = $UserClientRepository->findByClient($this->getUser());
+        $user = $UserClientRepository->findByClientAndUser($this->getUser(), $id);
+        if ($user === NULL){
+            return $this->json([
+                'status' => 404,
+                'message' => "No users have this id"
+            ],404);
+        }
         return $this->json($user, 200,[],['groups' => 'user:detail']);
     }
 
@@ -72,20 +86,11 @@ class UserController extends AbstractController
      * Insertion d'un utilisateur
      * @Route("/api/users", name="api_users_insert", methods="POST")
      * @TAG\Tag(name="Users")
-     * @TAG\Parameter(
-     *     name="firstname",
-     *     in="path",
+     * @TAG\RequestBody(
+     *     description="Add user",
      *     required=true,
-     *     description="Name of the user",
-     *     @TAG\Schema(type="string")
-     * )
-     * @TAG\Parameter(
-     *     name="lastname",
-     *     in="path",
-     *     required=true,
-     *     description="Name of the user",
-     *     @TAG\Schema(type="string")
-     * )
+     *     @Model(type=UserClient::class, groups={"user:detail"}))
+     *
      * @TAG\Response(
      *     response=201,
      *     description="Success - Return User added.",
@@ -100,15 +105,10 @@ class UserController extends AbstractController
      *     description="JWT Token not found | Expired JWT Token"
      * )
      * @TAG\Response(
-     *     response=409,
-     *     description="User already exists with this email for this company."
-     * )
-     * @TAG\Response(
      *     response=500,
      *     description="Internal Error"
      * )
      * @TAG\Tag(name="Users")
-     * @Security(name="Bearer")
      */
     public function insertUsers(Request $request,SerializerInterface $serializer,
                                 EntityManagerInterface $em, ValidatorInterface $validator)
@@ -151,7 +151,7 @@ class UserController extends AbstractController
      *     @TAG\Schema(type="integer")
      * )
      * @TAG\Response(
-     *     response=204,
+     *     response=200,
      *     description="Success, User deleted."
      * )
      * @TAG\Response(
@@ -166,23 +166,24 @@ class UserController extends AbstractController
      *     response=404,
      *     description="User not found."
      * )
+     *
+     * * @TAG\Response(
+     *     response=500,
+     *     description="Error 500, please contact ADMIN."
+     * )
+     *
      * @TAG\Tag(name="Users")
-     * @Security(name="Bearer")
      */
-    public function deleteUser(UserClient $user)
+    public function deleteUser(UserClient $user): JsonResponse
     {
-        try {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($user);
-            $em->flush();
-            return $this->json('Suppresion de l\'utilisateur effectué', 201);
-        } catch(NotEncodableValueException $e){
-            {
-                return $this->json([
-                    'status' => 400,
-                    'message' => $e->getMessage()
-                ],400);
-            }
+        if($this->getUser()->getId() !== $user->getClient()->getId()){
+            return $this->json(null,403);
         }
+
+        $em = $this->getDoctrine()->getManager();
+        $em->remove($user);
+        $em->flush();
+        return $this->json('User deleted successfully', 200);
+
     }
 }
